+++ src/pages/MenuDetail.vue
@@ -17,8 +17,8 @@
         <q-list separator class="q-pa-none">
           <q-item v-for="(modifier,index) in getRecsModifier" :key="index">
             <q-item-main class="text-grey-14">
-              <q-radio v-if="modifier.type==='size'" v-model="chosesize" :val="modifier" :label="modifier.name" color="brown-14" @input="onChangeExtra()" />
-              <q-checkbox v-else v-model="extra" :label="modifier.name" :val="modifier" color="brown-14" @input="onChangeExtra()" />
+              <q-radio v-if="modifier.type==='size'" v-model="sizes" :val="modifier" :label="modifier.name" color="brown-14" @input="onChangeModifier()" />
+              <q-checkbox v-else v-model="modifiers" :label="modifier.name" :val="modifier" color="brown-14" @input="onChangeModifier()" />
             </q-item-main>
             <q-item-side right class="text-grey-14">
               {{modifier.price|price}}
@@ -29,7 +29,7 @@
         <q-toolbar slot="footer" color="brown-6 row">
           <q-btn icon="add_shopping_cart" flat label="Add To Cart" />
           <q-toolbar-title class="text-right">
-            {{'$'+currentPrice.toFixed(2)}}
+            {{'$'+addToCartPrice.toFixed(2)}}
           </q-toolbar-title>
         </q-toolbar>
       </q-modal-layout>
@@ -44,9 +44,9 @@ export default {
   data() {
     return {
       opened: true,
-      chosesize: {},
-      currentPrice: 0,
-      extra: [],
+      sizes: {},
+      addToCartPrice: 0,
+      modifiers: [],
       menu: {},
     }
   },
@@ -78,16 +78,13 @@ export default {
       this.setCounter(0)
     },
     calculatePrice() {
-      var price = this.menu.price * this.getCounter
-      var extraprice =
-        _.sumBy(this.extra, ({price}) => {
-          return parseFloat(price)
-        }) * this.getCounter
-      var chosePrice = parseFloat(this.chosesize.price) * this.getCounter
-      return extraprice + price + chosePrice
+      var modifierPrice = _.sumBy(this.modifiers, ({price}) => {
+        return parseFloat(price)
+      })
+      return (modifierPrice + parseFloat(this.menu.price) + parseFloat(this.sizes.price || 0)) * this.getCounter
     },
-    onChangeExtra() {
-      this.currentPrice = this.calculatePrice()
+    onChangeModifier() {
+      this.addToCartPrice = this.calculatePrice()
     },
   },
   watch: {
@@ -95,14 +92,15 @@ export default {
       var modify = _.find(data, menu => {
         return menu.type === 'size' && menu.price === '0'
       })
-      this.chosesize = modify !== undefined ? modify : {}
+      this.sizes = modify !== undefined ? modify : {}
     },
     getCounter(val) {
-      this.currentPrice = this.calculatePrice()
+      this.addToCartPrice = this.calculatePrice()
     },
   },
   mounted() {
     this.menu = this.getRecs.find(item => item.id === parseFloat(this.$route.params.menuId))
+    this.addToCartPrice = this.calculatePrice()
     this.fetchModifiers(this.menu)
   },
 }